Removed stuff that was not needed
Changed the interface to abstract class
Remove GameMainController adn GameMain.fxml, not needed since the
GameMenu will be the landing view.
Added a swich scene method in Applications.
Checked rendering and exit button of GameMenu works.
Rest head to previous commit will be use f - push if this is needed